move config to file
